Switch to cmake build system
This was somewhat more complex than I thought it would be, but the
advantages are too many to ignore. Out of source builds are very
important.

The testing suite has been moved into a separate source file which is
included through its header `utils.h`. This header is generated by cmake
to include an absolute and up-to-date path to the files which are used
for some tests. Those files are copied to the build directory system
during make which makes the out-of-source build independent.